fix: some bug

+++ src/Api/DTOs/Scheduling/DriverData.cs
@@ -4,5 +4,5 @@ public class DriverData
 {
     public Guid Id { get; set; }
     public int SeatingCapacity { get; set; }
-    public int Used { get; set; }
+    public double Used { get; set; }
 }
